requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.1.0=pypi_0
appnope=0.1.3=pyhd8ed1ab_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py38h33210d7_2
asttokens=2.0.5=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
attrs=21.4.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pyha770c72_0
bleach=5.0.0=pyhd8ed1ab_0
brotli=1.0.9=h1c322ee_7
brotli-bin=1.0.9=h1c322ee_7
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2022.5.18.1=h4653dfc_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.2.0=pypi_0
certifi=2022.5.18.1=pypi_0
cffi=1.15.0=py38hc67bbb8_0
charset-normalizer=2.0.12=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.6.0=py38hea55cf5_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.3=pyhd8ed1ab_0
flatbuffers=1.12=pypi_0
flit-core=3.7.1=pyhd8ed1ab_0
fonttools=4.33.3=py38he5c2ac2_0
freetype=2.10.4=h17b34a0_1
gast=0.4.0=pypi_0
giflib=5.2.1=h27ca646_2
google-auth=2.6.6=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.46.3=py38h1ef021a_0
h5py=3.6.0=nompi_py38hacf61ce_100
hdf5=1.12.1=nompi_hd9dbc9e_104
idna=3.3=pypi_0
importlib-metadata=4.11.4=py38h10201cd_0
importlib_resources=5.7.1=pyhd8ed1ab_1
ipykernel=6.13.0=py38h2cb4d76_0
ipython=8.4.0=py38h10201cd_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.7.0=pyhd8ed1ab_0
jedi=0.18.1=py38h10201cd_1
jinja2=3.1.2=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h1c322ee_1
jsonschema=4.5.1=pyhd8ed1ab_0
jupyter=1.0.0=py38h10201cd_7
jupyter_client=7.3.1=pyhd8ed1ab_0
jupyter_console=6.4.3=pyhd8ed1ab_0
jupyter_core=4.10.0=py38h10201cd_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_widgets=1.1.0=pyhd8ed1ab_0
keras=2.9.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.2=py38h0f71543_1
krb5=1.19.3=he492e65_0
lcms2=2.12=had6a04f_0
lerc=3.0=hbdafb3b_0
libblas=3.9.0=14_osxarm64_openblas
libbrotlicommon=1.0.9=h1c322ee_7
libbrotlidec=1.0.9=h1c322ee_7
libbrotlienc=1.0.9=h1c322ee_7
libcblas=3.9.0=14_osxarm64_openblas
libclang=14.0.1=pypi_0
libcurl=7.83.1=h7965298_0
libcxx=14.0.4=h6a5c8ee_0
libdeflate=1.10=h3422bc3_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0.dev0=11_0_1_hf114ba7_23
libgfortran5=11.0.1.dev0=hf114ba7_23
liblapack=3.9.0=14_osxarm64_openblas
libnghttp2=1.47.0=hf30690b_0
libopenblas=0.3.20=openmp_h2209c59_0
libpng=1.6.37=hf7e6567_2
libsodium=1.0.18=h27ca646_1
libssh2=1.10.0=h7a5bd25_2
libtiff=4.4.0=h2810ee2_0
libwebp=1.2.2=h0d20362_0
libwebp-base=1.2.2=h3422bc3_1
libxcb=1.13=h9b22ae9_1004
libzlib=1.2.12=h90dfc92_0
llvm-openmp=14.0.4=hd125106_0
lz4-c=1.9.3=hbdafb3b_1
markdown=3.3.7=pypi_0
markupsafe=2.1.1=py38h33210d7_1
matplotlib=3.5.2=py38h150bfb4_0
matplotlib-base=3.5.2=py38hd7f6e49_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py38hea4295b_1005
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.5.13=pyhd8ed1ab_0
nbconvert=6.4.5=py38h10201cd_0
nbformat=5.4.0=pyhd8ed1ab_0
ncurses=6.3=h07bb92c_1
nest-asyncio=1.5.5=pyhd8ed1ab_0
notebook=6.4.11=pyha770c72_0
numpy=1.22.4=py38he1fcd3f_0
oauthlib=3.2.0=pypi_0
openjpeg=2.4.0=h062765e_1
openssl=3.0.3=ha287fd2_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=py38h6724fcd_2
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.1.1=py38h3852883_1
pip=22.1.2=pyhd8ed1ab_0
prometheus_client=0.14.1=pyhd8ed1ab_0
prompt-toolkit=3.0.29=pyha770c72_0
prompt_toolkit=3.0.29=hd8ed1ab_0
protobuf=3.19.4=pypi_0
psutil=5.9.1=py38he5c2ac2_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.12.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyrsistent=0.18.1=py38h33210d7_1
python=3.8.13=hd3575e6_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.15.3=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytz=2022.1=pyhd8ed1ab_0
pyyaml=6.0=py38h33210d7_4
pyzmq=23.0.0=py38hc61cae3_0
readline=8.1=hedafd6a_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-learn=1.1.1=py38hddc176c_0
scipy=1.8.1=py38h4f188a7_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=62.3.2=py38h10201cd_0
six=1.15.0=pypi_0
soupsieve=2.3.1=pyhd8ed1ab_0
sqlite=3.38.5=h40dfcc0_0
stack_data=0.2.0=pyhd8ed1ab_0
tensorboard=2.9.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-deps=2.9.0=0
tensorflow-estimator=2.9.0=pypi_0
tensorflow-macos=2.9.1=pypi_0
tensorflow-metal=0.5.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.15.0=py38h10201cd_0
testpath=0.6.0=pyhd8ed1ab_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=he1e0b03_0
tornado=6.1=py38h33210d7_3
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.2.2.post1=pyhd8ed1ab_0
typing-extensions=4.2.0=pypi_0
unicodedata2=14.0.0=py38h33210d7_1
urllib3=1.26.9=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=3.6.0=py38h10201cd_0
wrapt=1.14.1=pypi_0
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.5=h642e427_1
yaml=0.2.5=h3422bc3_2
zeromq=4.3.4=hbdafb3b_1
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.12=h90dfc92_0
zstd=1.5.2=hd705a24_1
